303 work.
This is the work completed in the Spring 2021 semester. While not perfect, it does a good job of finding the paths optimally considering the cover and obstacle logic implemented. This semester, I will work on making the UI better, and on perfecting the code to abstract some of the work, clean it up, etc. The cover logic will also need to be edited in unison with the ML aspect of the project.